requirements.txt
PyAudio
wave
sqlalchemy
llama-index
mysql-connector-python
SpeechRecognition
openai-whisper
soundfile
edge_tts
pygame
keyboard